refactor: move config to cluster to reflect usage in cluster comms
